+++ grafana-plugin/src/pages/incident/Incident.tsx
@@ -126,7 +126,7 @@ class IncidentPage extends React.Component<IncidentPageProps, IncidentPageState>
     if (!incident && !isNotFoundError && !isWrongTeamError) {
       return (
         <div className={cx('root')}>
-          <LoadingPlaceholder text="Loading alert group..." />
+          <LoadingPlaceholder text="Loading Alert Group..." />
         </div>
       );
     }
@@ -139,10 +139,10 @@ class IncidentPage extends React.Component<IncidentPageProps, IncidentPageState>
               <div className={cx('not-found')}>
                 <VerticalGroup spacing="lg" align="center">
                   <Text.Title level={1}>404</Text.Title>
-                  <Text.Title level={4}>Incident not found</Text.Title>
+                  <Text.Title level={4}>Alert group not found</Text.Title>
                   <PluginLink query={{ page: 'incidents', cursor, start, perpage }}>
                     <Button variant="secondary" icon="arrow-left" size="md">
-                      Go to incidents page
+                      Go to Alert Groups page
                     </Button>
                   </PluginLink>
                 </VerticalGroup>
@@ -250,7 +250,7 @@ class IncidentPage extends React.Component<IncidentPageProps, IncidentPageState>
                   <IconButton
                     name="share-alt"
                     onClick={this.showAttachIncidentForm}
-                    tooltip="Attach to another alert group"
+                    tooltip="Attach to another Alert Group"
                     className={cx('title-icon')}
                   />
                 )}
@@ -395,7 +395,10 @@ class IncidentPage extends React.Component<IncidentPageProps, IncidentPageState>
             </li>
           ))}
         </ul>
-        <Field label="Leave a resolution note" description="Will also show up in the thread of incident in Slack">
+        <Field
+          label="Leave a resolution note"
+          description="Will also show up in the thread of the Alert Group in Slack"
+        >
           <TextArea
             value={resolutionNoteText}
             onChange={(e: any) => this.setState({ resolutionNoteText: e.target.value })}